--- src/components/alertMessage/__docs__/AlertMessage.mdx
@@ -1,39 +1,0 @@
-import React from "react";
-import { Meta, Canvas, Controls } from "@storybook/blocks";
-import * as AlertMessage from "./AlertMessage.stories.tsx";
-
-<Meta of={AlertMessage} title="AlertMessage" />
-
-#### Example
-
-<Canvas of={AlertMessage.Positive} />
-
-# AlertMessageExample
-
-
-Example for usage of `AlertMessage` component.
-
-## Usage
-
-```tsx
-import AlertMessage from "./AlertMessage";
-
-const alertMessageExampleProps = {
-  alertMessage: "Your changes have been saved successfully.",
-  alertLink: "/details",
-  linkText: "View details",
-  alertType: "positive",
-};
-
-const AlertMessageExample = () => (
-  <div style={{ width: "20rem", color: "#115873" }}>
-    <AlertMessage {...alertMessageExampleProps} />
-  </div>
-);
-
-export default AlertMessageExample;
-
-```
-
-## Props
-<Controls of={AlertMessage.Positive} />
