--- src/components/alertMessage/index.ts
@@ -1,1 +1,0 @@
-export { default as AlertMessage } from "./AlertMessage";
